+++ actix-http/src/http_message.rs
@@ -44,15 +44,15 @@ pub trait HttpMessage: Sized {
         }
     }
 
-    /// Read the request content type. If request did not contain a *Content-Type* header, an empty
-    /// string is returned.
-    fn content_type(&self) -> &str {
-        if let Some(content_type) = self.headers().get(header::CONTENT_TYPE) {
-            if let Ok(content_type) = content_type.to_str() {
-                return content_type.split(';').next().unwrap().trim();
-            }
-        }
-        ""
+    /// Read the request content type.
+    ///
+    /// If request does not contain a *Content-Type* header, or if the header value is not valid
+    /// UTF-8, `None` is returned.
+    fn content_type(&self) -> Option<&str> {
+        self.headers()
+            .get(header::CONTENT_TYPE)
+            .and_then(|ct| ct.to_str().ok())
+            .map(|ct| ct.split_once(';').map_or(ct, |(ct, _)| ct).trim())
     }
 
     /// Get content type encoding.
@@ -142,13 +142,20 @@ mod tests {
         let req = TestRequest::default()
             .insert_header(("content-type", "text/plain"))
             .finish();
-        assert_eq!(req.content_type(), "text/plain");
+        assert_eq!(req.content_type(), Some("text/plain"));
         let req = TestRequest::default()
             .insert_header(("content-type", "application/json; charset=utf-8"))
             .finish();
-        assert_eq!(req.content_type(), "application/json");
+        assert_eq!(req.content_type(), Some("application/json"));
         let req = TestRequest::default().finish();
-        assert_eq!(req.content_type(), "");
+        assert_eq!(req.content_type(), None);
+        let req = TestRequest::default()
+            .insert_header((
+                header::CONTENT_TYPE,
+                header::HeaderValue::from_bytes(b"\xff").unwrap(),
+            ))
+            .finish();
+        assert_eq!(req.content_type(), None);
     }
 
     #[test]